pyproject.toml
[project]
name = "devcontainer"
version = "0.1.0"
description = "Personal devcontainer for various projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pyyaml>=6.0.2"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "podman-compose>=1.3.0",
]
